lab/requirements.txt
apache-airflow-providers-apache-spark==3.0.0
pyspark==3.3.0
delta-spark==2.1.0
pandas==1.5.0
yfinance==0.1.74
requests==2.28.1
airflow-pentaho-plugin==1.0.16

#https://pypi.org/search/?q=apache-airflow-providers&o=
apache-airflow-providers-microsoft-azure==4.3.0
apache-airflow-providers-microsoft-mssql==3.2.1
apache-airflow-providers-cncf-kubernetes==4.4.0
apache-airflow-providers-databricks==3.3.0
apache-airflow-providers-snowflake==3.3.0

# https://pypi.org/project/astro-sdk-python/
astro-sdk-python==1.2.3

# https://pypi.org/project/astronomer-providers/
astronomer-providers==1.11.2

# https://pypi.org/project/minio/
minio==7.1.12